+++ test/unit/mysql_db_admin/detect_dbs.py
@@ -41,10 +41,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:
-
     Methods:
         setUp -> Initialize testing environment.
         test_no_miss_db -> Test with no missing databases.