+++ test/unit/sonarqube_code_coverage.sh
@@ -15,6 +15,9 @@ coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/checksum.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/detect_dbs.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/optimize.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/process_request.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/proc_all_dbs.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/proc_all_tbls.py
+coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/proc_some_tbls.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_check.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_checksum.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_optimize.py
@@ -22,7 +25,6 @@ coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_analyze.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/main.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/run_program.py
 coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/status.py
-coverage run -a --source=mysql_db_admin test/unit/mysql_db_admin/setup_mail.py
 
 echo ""
 echo "Producing code coverage report"