+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.1.0] - 2019-12-06
+### Fixed
+- run_analyze:  Fixed problem with mutable default arguments issue.
+- run_optimize:  Fixed problem with mutable default arguments issue.
+
+### Added
+- Added a program lock mechanism to the program.
+- Added -z option for standard out suppression.
+
+### Changed
+- main:  Added gen_class.ProgramLock class and code check.
+- run_program:  Replaced setup_mail call with gen_class.setup_mail call.
+- status:  Added print JSON format to standard out, unless standard out is suppressed.
+- Documentation update.
+
+### Removed
+- setup_mail:  No longer required, replaced with call to gen_class.setup_mail.
+
+
 ## [3.0.0] - 2019-07-16
 Breaking Change